% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function produces a map of accidents in a single state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{is an integer indicating which state's data will be
filtered and then mapped. Parameter can be any integer 1-56.}

\item{year}{is a single integer representing years reported on by
the Fatality Analysis Reporting System. Parameter can years 2013-2015.}
}
\value{
This function returns a map showing the location of accidents in
a single state from a single year based on data from the the Fatality
Analysis Reporting System.
}
\description{
This function produces a map of accidents in a single state
}
\examples{
acc_years <- fars_map_state(10, 2015)

}
