% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads CSV file into R}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A file path of a CSV file to be converted}
}
\value{
This function returns a tibble dataframe from a CSV file.
If the file path is wrong or the file does not exist then an error message
will return.
}
\description{
This is a function that takes a filename as a parameter which is supposed
to be a CSV file and reads that file into R and then converts it to a
tibble dataframe.
}
\examples{
acc_2015 <- fars_read("path/accident_2015.csv")
acc_2014 <- fars_read("path/accident_2014.csv")
acc_2013 <- fars_read("path/accident_2013.csv")

}
