% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This function return a list of dataframes}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{is a vector of integers or single integer representing
years reported on by the Fatality Analysis Reporting System.}
}
\value{
This function returns a list of tibble dataframes containing
the month and year varibles from the accident files from the
Fatality Analysis Reporting System. The dataframes are compliled
based on the years specified in a vector from the 'years' parameter.
If only one year is provided that a simple tibble dataframe is returned.
}
\description{
Applied accross multiple or single number of CSV files based on what
is specified in the 'years' paramater; this function reads in CSV file(s)
and converts them into a tibble dataframe and then selects only the 'MONTH'
and 'year' columns.
}
\examples{
acc_years <- fars_read_years(c(2013, 2014, 2015))

}
