% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This function summarizes a dataframe from CSV file}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{is a vector of integers or single integer representing
years reported on by the Fatality Analysis Reporting System.}
}
\value{
This function returns a dataframe that summarizes the three
dataframe objects from the list created by the 'fars_read_years' function.
The dataframe contains four columns and twelve rows. The 'MONTH' column
has the month number and the three year columns contain the number of
fatalities in the corresponding month.
}
\description{
This function summarizes the number of accidents reported by the Fatality
Analysis Reporting System by year and month.
}
\examples{
acc_years <- fars_summarize_years(c(2013, 2014, 2015))

}
