% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates filename for CSV file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{is an integer that is a year reported by the Fatality Analysis Reporting System}
}
\value{
This function returns a tibble filepath for a CSV file containing data for the
Fatality Analysis Reporting System
}
\description{
This is a function that takes a year parameter and creates a filename for
Fatality Analysis Reporting System data.
}
\examples{
make_filename(2013)
make_filename(2014)
make_filename(2015)

}
