% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw a dot map of FARS data for a particular year in a particular state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer representing a US state number.
Must be a state number which appears in the data file for the relevant year, otherwise
an error message will result.}

\item{year}{A 4-digit number denoting a year.
Must be a year for which you have a FARS data file, otherwise an error message will result.}
}
\value{
A map of the accidents occuring that year in that state.
}
\description{
This function should be supplied with a state number and a 4-digit year, and will draw
a dot map of all the accidents occurring that year in that state.
}
\examples{
fars_map_state(40,2013)
}
