% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a file of FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name of the file you want to read.
The file must exist, otherwise an error message will result.}
}
\value{
A tibble of the file contents
}
\description{
This function reads the FARS data from a specified file, and returns it as a tibble.
}
\details{
NB - in practice, this function is only likely to be called from within other functions.
However, it is exported and documented here for the sake of completeness.
}
\examples{
fars_read(make_filename(2013))
}
