% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read all the FARS data for specified years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of 4-digit numbers, each denoting a year.
Must all be years for which you have a FARS data file, otherwise an error message will result.}
}
\value{
A list of tibbles, one for each of the specified \code{years},
retaining the 'MONTH' and 'year' columns only.
}
\description{
This function should be supplied with a vector of 4-digit years. It will read the FARS
data from all the relevant files for those years, and return one tibble per year, retaining
the 'MONTH' and 'year' columns only.
}
\details{
NB - In practice, this function is only likely to be called from within other functions.
However, it is exported and documented here for the sake of completeness.
}
\examples{
fars_read_years(c(2013,2014))
}
