% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars2.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Function `fars_read_years`: Reads years and returns list with corresponding months and years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{It can be a vector or a list of number representing years with
format 'yyyy', such as in '1927'}
}
\value{
If 'years' has element corresponding with data bases's years
a list of months corresponding to the/these year/s is given back
other wise a exception error is thrown.
}
\description{
This function reads a vector or list 'years' with four digit
number in the format yyyy (e.g., 1927) and if the years exist/are correct
the function wil returns a list with the correspondung month and year found
in the data
}
\examples{
fars_read_years(c(2013, 2015))
fars_read_years(2013:2015)
fars_read_years(list(2013,2015))

}
