% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates a cross-tab with number of accidents per month and year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of 4-digit numbers, each denoting a year.
Must all be years for which you have a FARS data file, otherwise an error message will result.}
}
\value{
A tibble containing the number of accidents for each combination of month and year.
}
\description{
This function should be supplied with a vector of 4-digit years. It will read the FARS
data from all the relevant files for those years, and return a cross-tabulation showing
how many accidents there were in each month of each year.
}
\examples{
fars_summarize_years(c(2013,2014))
}
