% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars2.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Function `fars_summarize_years` Takes years in format yyyy and returns the number
of observations per month per year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{It can be a vector or a list of number representing years with
format 'yyyy', such as in '1927'}
}
\value{
a dplyr tibble object with the number of observations per
 month per year if the later are found to be describe in the data objects.
 Months are displayed as a column and years as colnames; number of
 observations are printed under years corresponding by months
}
\description{
Function `fars_summarize_years` Takes years in format yyyy and returns the number
of observations per month per year
}
\examples{
fars_summarize_years(years = c(2013, 2015))
fars_summarize_years(years = 2013:2015)
fars_summarize_years(years = list(2013,2015))

}
