% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct a file name for the FARS data for a particular year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A 4-digit number denoting a year.}
}
\value{
A string representing the name of the file containing FARS data for \code{year}
}
\description{
This function should be supplied with a 4-digit year, and will return
the name of the file in which FARS data for that year can be found.
}
\details{
NB - in practice, this function is only likely to be called from within other functions.
However, it is exported and documented here for the sake of completeness.
}
\examples{
make_filename(2013)
}
