% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars2.R
\name{make_filename}
\alias{make_filename}
\title{Function `make_file`: Make a valid filename given a year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a string or integer of four digits indicating the wanted year}
}
\value{
as a side effect, this functions prints a string with a valid
fars filename such as 'accident_1997.csv.bz2'
}
\description{
This function takes a year in the format yyyy (e.g 1977, 2007, etc)
and it gives back a valid fars filename.
}
\examples{
make_filename(2001)
make_filename(1987)

}
